requirements.txt
# Core dependencies
anduril-lattice-sdk>=1.0.0
pyyaml>=6.0
click>=8.1.0
python-dateutil>=2.8.2

# Geographic calculations
geopy>=2.3.0
pyproj>=3.6.0

# Math and calculations
numpy>=1.24.0

# Logging (using standard library, but structlog optional)
structlog>=23.1.0

# Type checking
mypy>=1.5.0
types-pyyaml>=6.0.12
types-python-dateutil>=2.8.19

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.7.0
flake8>=6.1.0

